refactor(drasi): split DrasiReactiveGraph.tsx into subcomponents
Splits the 2885-line monolith into 9 focused modules under
web/src/components/cards/drasi/:

  DrasiTypes.ts          — all shared interfaces and type aliases
  DrasiConstants.ts      — every named constant and palette value
  DrasiDemoData.ts       — 5 themed demo pipelines + row generators
  DrasiFlowUtils.ts      — union-find flow discovery (computeFlows)
  DrasiNodeCard.tsx      — NodeCard, NodeControls, StatusDot, icon helpers
  DrasiFlowLine.tsx      — FlowLine SVG component with animated dots
  DrasiResultsTable.tsx  — ResultsTable, KPIBox, sort/format helpers
  DrasiModals.tsx        — ModalShell, ExpandModal, SourceConfigModal,
                           QueryConfigModal, ConnectionsModal, RowDetailDrawer
  DrasiStreamSamples.tsx — StreamSampleDrawer + 6-language STREAM_SAMPLES

DrasiReactiveGraph.tsx is now the main composition root (1132 lines).
The public export via index.ts is unchanged.

Fixes #8624
